+++ test/Helper/HeadLinkTest.php
@@ -445,4 +445,10 @@ class HeadLinkTest extends \PHPUnit_Framework_TestCase
         $this->helper->appendStylesheet(['rel' => 'icon', 'href' => '/bar/baz', 'sizes' => '123x456']);
         $this->assertContains('sizes="123x456"', $this->helper->toString());
     }
+    
+    public function testItempropAttributeIsSupported()
+    {
+        $this->helper->prependAlternate(['itemprop' => 'url', 'href' => '/bar/baz', 'rel' => 'canonical']);
+        $this->assertContains('itemprop="url"', $this->helper->toString());
+    }
 }